pyproject.toml
[tool.poetry]
name = "prodtech-attributes"
version = "0.1.0"
description = "Repository for storing standardized attribute identifiers for objects relevant in modeling production- and injection networks"
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
pyyaml = "^6.0.2"
pydantic-extra-types = "^2.10.1"


[tool.poetry.group.test.dependencies]
pytest = "^8.3.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
